+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 4.3.0 (2023-03-29)
 
 - The names of staked/locked/earned tokens are differentiated from raw tokens, to avoid clobbering. #194
-- Add LDO and agEUR tokens
+- Add LDO, SUSHI and agEUR tokens
 
 ## 4.2.14 (2023-03-27)